+++ packages/cap-config-writer/src/cap-config/types.ts
@@ -1,4 +1,4 @@
-import type { CapService } from '@sap-ux/odata-service-inquirer';
+import type { CdsVersionInfo } from '@sap-ux/project-access';
 
 export type CdsUi5PluginInfo = {
     /**
@@ -18,6 +18,40 @@ export type CdsUi5PluginInfo = {
      */
     hasCdsUi5Plugin: boolean;
 };
+
+export type CapRuntime = 'Node.js' | 'Java';
+
+export interface CapService {
+    /**
+     * The path to the CAP project.
+     */
+    projectPath: string;
+    /**
+     * The CDS info for the Cds instance that was used to compile the project when determining the service.
+     */
+    cdsVersionInfo?: CdsVersionInfo;
+    /**
+     * The name of the CAP service as identified by the cds model.
+     */
+    serviceName: string;
+    /**
+     * The URL path to the service, as specfied in the manifest.json of generated apps
+     * This is also provided as `OdataServicePromptAnswers` property `servicePath`
+     */
+    urlPath?: string;
+    /**
+     * The relative path (from the `projectPath`) to the service cds file.
+     */
+    serviceCdsPath?: string;
+    /**
+     * The runtime of the Cds instance that was used to compile the project when determining the service.
+     */
+    capType?: CapRuntime;
+    /**
+     * The relative path (from the `projectPath`) to the app folder
+     */
+    appPath?: string;
+}
 
 export interface CapServiceCdsInfo extends CapService {
     cdsUi5PluginInfo: CdsUi5PluginInfo;