+++ packages/cards-editor-config-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/cards-editor-config-writer
 
+## 0.4.9
+
+### Patch Changes
+
+-   Updated dependencies [19aad96]
+    -   @sap-ux/ui5-config@0.26.1
+
 ## 0.4.8
 
 ### Patch Changes